Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -7,13 +7,13 @@ int main() {
     {
         temp*=2;
     }
-    int ans=0;
+    long long  ans=0;
     while(temp!=1){
         int q =num/temp;
         ans =ans*10+q;
         num= num-(q*temp);
         temp=temp/2;
     }
-    printf("%d",ans);
+    printf("%lld",ans);
     return 0;
 }
